Avoid scrolling if user has scrolled up
Messages list view will only scroll to the bottom if the user hasn't
scrolled up. If the scrollbar is set to the extreme bottom, it auto
scrolls like it should. This is a major UX improvement that makes it
less painful to read old messages as newer messages come through.